Change TransformableVideoSenderFrame::GetTimestamp to return capture time
Bug: None

+++ api/frame_transformer_interface.h
@@ -35,7 +35,7 @@ class TransformableFrameInterface {
 
   virtual uint8_t GetPayloadType() const = 0;
   virtual uint32_t GetSsrc() const = 0;
-  virtual uint32_t GetTimestamp() const = 0;
+  virtual int64_t GetTimestamp() const = 0;
 
   enum class Direction {
     kUnknown,
