+++ video/rtp_video_stream_receiver_frame_transformer_delegate.cc
@@ -43,7 +43,7 @@ class TransformableVideoReceiverFrame
 
   uint8_t GetPayloadType() const override { return frame_->PayloadType(); }
   uint32_t GetSsrc() const override { return ssrc_; }
-  uint32_t GetTimestamp() const override { return frame_->Timestamp(); }
+  int64_t GetTimestamp() const override { return frame_->Timestamp(); }
 
   bool IsKeyFrame() const override {
     return frame_->FrameType() == VideoFrameType::kVideoFrameKey;
